requirements.txt
openai-whisper==20231117
langchain==0.2.15
langchain_chroma==0.1.3
langchain_community==0.2.14
langchain-ollama==0.1.1
yt-dlp==2024.8.6
chromadb==0.5.5
dynaconf==3.2.6
